Fixed race condition
The two phase transactionality algorithm only supports failing if the current state of the transaction is "Prepare", therefore, if any of the entitties fail when in the "Commit" phase, the state is undefined. This implementation does not support this. The race condition was given by a timeout in the second phase (entities failed to send a commit message in the "commit" phase in time). This time restriction in the second phase was removed.